requirements.txt
aiogram>=3.13,<4.0
python-dotenv>=1.0.1,<2.0
aiohttp>=3.10,<4.0
yt-dlp>=2025.10.14
